Bump json gem to 1.8.3 so that gh-pages can be generated with ruby 2.2.x
When installing the json gem with ruby 2.2.x, compilation fails. The
details of why it is failing are in this[0] github issue. Subsequent gem
releases (1.8.2+) have fixed the issue, so this commit bumps the gem to
the latest non-broken version so that people with ruby 2.2.x can install
and run the gh-pages for rails-erd locally.

[0] - https://github.com/flori/json/issues/229