Fix too broad regular expression in test matching almost everything
Escape the literal pipe character `|` in regular expressions.

The pipe character, when not escaped, is used in regular expressions to enumerate alternatives. The regex `/cat|dog/` matches both the string "cat" as well as the string "dog".

The regex with the unescaped pipe character in the test "node record labels should not have direction reversing braces for horizontal orientation" matches any string that ends with a double quotation mark `"` and does not actually test whether the test input contains braces.

After this change all four tests dealing with direction reversing braces fail as expected.