Fix usage of orientation option in record templates
Depending on the graphs orientation, record nodes need to be wrapped in curly braces `{ }` so that the inner layout orientation of the record is always top-to-bottom and does not change with the global graph orientation.
Since #183 swapped the meaning of the :vertical and :horizontal orientation options, the templates need to be changed accordingly. This closes #218.