Pin minitest to 5.10.1
I got back into the project this morning and noticed a TON of test
failures in recent PRs. 5 min of research led me to these issues in
recent versions of minitest and Rails. Since we use a variety of Rails
versions, pinning the version minitest to one that works across legacy
Rails versions is a solution for today, but we may have to take a look
at a more sophisticated solution in the future.

seattlerb/minitest#730
rails/rails#31624