Make loading environment.rb optional
This environment file is the only thing that makes it impossible to use rails-erd with an
application which uses ActiveRecord but is not a Rails application. Since the objective of
loading that file is to eager-load the models, we can proceed without having loaded it - as
long as the user is aware.